Simplify simple task handlers
