+++ concrete/src/Permission/Category/TaskHandler/Sitemap.php
@@ -2,77 +2,24 @@
 
 namespace Concrete\Core\Permission\Category\TaskHandler;
 
-use Concrete\Core\Controller\Controller;
 use Concrete\Core\Error\UserMessageException;
-use Concrete\Core\Http\ResponseFactoryInterface;
-use Concrete\Core\Permission\Access\Access;
-use Concrete\Core\Permission\Access\Entity\Entity;
-use Concrete\Core\Permission\Category\TaskHandlerInterface;
+use Concrete\Core\Permission\Category\DefaultTaskHandler;
 use Concrete\Core\Permission\Checker;
-use Concrete\Core\Permission\Duration;
-use Concrete\Core\Permission\Key\SitemapKey;
-use Symfony\Component\HttpFoundation\Response;
 
 defined('C5_EXECUTE') or die('Access Denied.');
 
-class Sitemap extends Controller implements TaskHandlerInterface
+class Sitemap extends DefaultTaskHandler
 {
     /**
      * {@inheritdoc}
      *
-     * @see \Concrete\Core\Permission\Category\TaskHandlerInterface::handle()
+     * @see \Concrete\Core\Permission\Category\DefaultTaskHandler::checkAccess()
      */
-    public function handle(string $task, array $options): ?Response
+    protected function checkAccess(): void
     {
         $p = new Checker();
         if (!$p->canAccessTaskPermissions()) {
             throw new UserMessageException(t('Access Denied.'));
         }
-        $method = lcfirst(camelcase($task));
-        if (!method_exists($this, $method)) {
-            throw new UserMessageException(t('Unknown permission task: %s', $task));
-        }
-
-        return $this->{$method}($options);
-    }
-
-    protected function addAccessEntity(array $options): ?Response
-    {
-        $pk = SitemapKey::getByID($options['pkID']);
-        $pa = Access::getByID($options['paID'], $pk);
-        $pe = Entity::getByID($options['peID']);
-        $pd = empty($options['pdID']) ? null : Duration::getByID($options['pdID']);
-        $pa->addListITem($pe, $pd, $options['accessType']);
-
-        return $this->app->make(ResponseFactoryInterface::class)->json(true);
-    }
-
-    protected function removeAccessEntity(array $options): ?Response
-    {
-        $pk = SitemapKey::getByID($options['pkID']);
-        $pa = Access::getByID($options['paID'], $pk);
-        $pe = Entity::getByID($options['peID']);
-        $pa->removeListItem($pe);
-
-        return $this->app->make(ResponseFactoryInterface::class)->json(true);
-    }
-
-    protected function savePermission(array $options): ?Response
-    {
-        $pk = SitemapKey::getByID($options['pkID']);
-        $pa = Access::getByID($options['paID'], $pk);
-        $pa->save($options);
-
-        return $this->app->make(ResponseFactoryInterface::class)->json(true);
-    }
-
-    protected function displayAccessCell(array $options): ?Response
-    {
-        $pk = SitemapKey::getByID($options['pkID']);
-        $this->set('pk', $pk);
-        $this->set('pa', Access::getByID($options['paID'], $pk));
-        $this->setViewPath('/backend/permissions/labels');
-
-        return null;
     }
 }
