+++ concrete/src/Permission/Category/TaskHandler/User.php
@@ -2,65 +2,40 @@
 
 namespace Concrete\Core\Permission\Category\TaskHandler;
 
-use Concrete\Core\Controller\Controller;
 use Concrete\Core\Error\UserMessageException;
 use Concrete\Core\Http\ResponseFactoryInterface;
 use Concrete\Core\Permission\Access\Access;
-use Concrete\Core\Permission\Access\Entity\Entity;
-use Concrete\Core\Permission\Category\TaskHandlerInterface;
+use Concrete\Core\Permission\Category\DefaultTaskHandler;
 use Concrete\Core\Permission\Checker;
-use Concrete\Core\Permission\Duration;
-use Concrete\Core\Permission\Key\UserKey;
+use Concrete\Core\Permission\Key\Key;
 use Concrete\Core\Workflow\Workflow;
 use Symfony\Component\HttpFoundation\Response;
 
 defined('C5_EXECUTE') or die('Access Denied.');
 
-class User extends Controller implements TaskHandlerInterface
+class User extends DefaultTaskHandler
 {
     /**
      * {@inheritdoc}
      *
-     * @see \Concrete\Core\Permission\Category\TaskHandlerInterface::handle()
+     * @see \Concrete\Core\Permission\Category\DefaultTaskHandler::checkAccess()
      */
-    public function handle(string $task, array $options): ?Response
+    protected function checkAccess(): void
     {
         $p = new Checker();
         if (!$p->canAccessTaskPermissions()) {
             throw new UserMessageException(t('Access Denied.'));
         }
-        $method = lcfirst(camelcase($task));
-        if (!method_exists($this, $method)) {
-            throw new UserMessageException(t('Unknown permission task: %s', $task));
-        }
-
-        return $this->{$method}($options);
     }
 
-    protected function addAccessEntity(array $options): ?Response
-    {
-        $pk = UserKey::getByID($options['pkID']);
-        $pa = Access::getByID($options['paID'], $pk);
-        $pe = Entity::getByID($options['peID']);
-        $pd = empty($options['pdID']) ? null : Duration::getByID($options['pdID']);
-        $pa->addListItem($pe, $pd, $options['accessType']);
-
-        return $this->app->make(ResponseFactoryInterface::class)->json(true);
-    }
-
-    protected function removeAccessEntity(array $options): ?Response
-    {
-        $pk = UserKey::getByID($options['pkID']);
-        $pa = Access::getByID($options['paID'], $pk);
-        $pe = Entity::getByID($options['peID']);
-        $pa->removeListItem($pe);
-
-        return $this->app->make(ResponseFactoryInterface::class)->json(true);
-    }
-
+    /**
+     * {@inheritdoc}
+     *
+     * @see \Concrete\Core\Permission\Category\DefaultTaskHandler::savePermission()
+     */
     protected function savePermission(array $options): ?Response
     {
-        $pk = UserKey::getByID($options['pkID']);
+        $pk = Key::getByID($options['pkID']);
         $pa = Access::getByID($options['paID'], $pk);
         $pa->save($options);
         $pa->clearWorkflows();
@@ -76,19 +51,9 @@ class User extends Controller implements TaskHandlerInterface
         return $this->app->make(ResponseFactoryInterface::class)->json(true);
     }
 
-    protected function displayAccessCell(array $options): ?Response
-    {
-        $pk = UserKey::getByID($options['pkID']);
-        $this->set('pk', $pk);
-        $this->set('pa', Access::getByID($options['paID'], $pk));
-        $this->setViewPath('/backend/permissions/labels');
-
-        return null;
-    }
-
     protected function saveWorkflows(array $options): ?Response
     {
-        $pk = UserKey::getByID($options['pkID']);
+        $pk = Key::getByID($options['pkID']);
         $pk->clearWorkflows();
         if (is_array($options['wfID'] ?? null)) {
             foreach ($options['wfID'] as $wfID) {
